requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
pyaml
dill
#-e .